requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
click==7.1.2
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
gTTS==2.2.3
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.19.5
packaging==21.0
pathy==0.6.0
preshed==3.0.5
PyAudio==0.2.11
pydantic==1.7.4
pyparsing==2.4.7
requests==2.25.1
six==1.15.0
smart-open==5.1.0
spacy==3.0.0
spacy-legacy==3.0.8
SpeechRecognition==3.8.1
srsly==2.4.1
thinc==8.0.8
tqdm==4.61.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
django==5.0.4
transformers==4.39.3
pytorch==2.4.0
